added branch module5-task1 and divided code into modules
